Updating the Documentation for SacSOS
Please feel free to give me feedback to improve the overall
documentation here.  The mobile readme.md have one of the best
directions for someone starting out, but I didn’t want to just copy
that over to main documentation without feedback from the community to
what to add and what to exclude.